#3240 Manually fixed multiple flaws with the automatic upgrade.

+++ app/Service/Controller/Api/V1/APIDungeonRouteControllerService.php
@@ -22,7 +22,7 @@ class APIDungeonRouteControllerService implements APIDungeonRouteControllerServi
         ?int         $viewportHeight = null,
         ?int         $imageWidth = null,
         ?int         $imageHeight = null,
-        ?int         $zoomLevel = null,
+        ?float       $zoomLevel = null,
         ?int         $quality = null,
     ): Collection {
         return $this->thumbnailService->queueThumbnailRefreshForApi(
